refactor(notifications): split notification_service.py (1277 LOC) into dispatcher + per-event flows (#73)
notification_service.py: 1277 LOC, 30-method monolith → 126 LOC thin
facade. Splits the single class into a dispatcher (persistence + user
preferences) plus per-event flow classes, all under
`core/services/notifications/`.

NEW package `core/services/notifications/`:
- `dispatcher.py` (460 LOC) — `NotificationDispatcher` class with the 11
  persistence methods (preferences, in-app notifications, email log
  CRUD, cleanup, send_email_notification).
- `org_invitation_flow.py` (273 LOC) — `OrgInvitationFlow` class with
  notify_organization_invitation / notify_invitation_accepted /
  notify_invitation_declined.
- `membership_change_flow.py` (394 LOC) — `MembershipChangeFlow` class
  with notify_membership_added / notify_role_changed /
  notify_membership_removed. The 3 nested `_update`/`_send` closures
  from the old monolith are promoted to private methods on this class.
- `beta_access_flow.py` (216 LOC) — `BetaAccessFlow` class with the 5
  beta-access notify methods. The repeated iscoroutinefunction /
  isawaitable / asyncio.run scaffolding is consolidated into a single
  `_send_direct_email` helper.
- `constants.py` (25 LOC) — shared EVENT_* and TEMPLATE_* constants
  (extracted to break a circular import between the facade and the
  dispatcher / flow classes).
- `__init__.py` (57 LOC) — package marker that re-exports the classes
  and constants for downstream imports.

`core/services/notification_service.py` (FACADE, 126 LOC):
- Same `NotificationService` class signature as before — public API
  fully preserved for the ~7 caller files (core/api/notifications.py,
  core/api/orgs.py, core/api/support.py, core/services/beta_access_service.py,
  3 test files).
- Constructor composes the dispatcher + 3 flow classes.
- Each public method is a 1-liner delegating to the dispatcher or the
  appropriate flow class.
- Re-exports the EVENT_* and TEMPLATE_* constants so legacy imports
  like `from core.services.notification_service import EVENT_ORG_INVITATION`
  keep working.

Each flow holds a reference to the FACADE (not just the dispatcher) so
that test patches on `service.send_email_notification = _fake_send`
propagate at call time. Without this indirection, tests that
intercept-via-attribute-replacement break because the flow's
`self._dispatcher.send_email_notification` is a different attribute.

Test-fixture fix bundled:
- tests/conftest.py: `db_session` fixture now declares an explicit
  `_migrated_db` dependency. Previously db_session depended on _engine
  (which depends on _test_postgres) but NOT on _migrated_db (the
  alembic upgrade fixture). Pytest's autouse session-scope resolution
  order is fragile when two fixtures share a transitive dep; this
  refactor's import shape happened to surface the latent ordering
  issue. The explicit dependency guarantees migrations run before any
  test acquires a connection. Confirmed by running
  test_notify_organization_invitation_role_is_string in isolation
  (passes now, was failing with `relation "users" does not exist`).

Coverage:
- Added `core/services/notifications/*` to pyproject.toml's
  [tool.coverage.run].omit list. The flow classes are extracted email/
  external-integration code; they share the rationale of the existing
  notification_service.py and email_service.py omit entries
  ("Email orchestration layers integrate with third-party providers
  and are covered by integration tests").

Net diff: 9 files changed, +1526/-1244 (net +282 LOC; the increase is
per-file class boilerplate — necessary cost of separating concerns).

Verification (run locally; integration branch is CI-silent):
- uv run pytest (full suite): 859 passed, 8 skipped, 80.26% coverage
- Facade preserves all 30 public methods + the EVENT/TEMPLATE constants
  via re-exports
- No nested `def _update`/`def _send` closures in any flow file (the
  smell scan flagged these; verified via rg)

Refs #73, umbrella #85.
