+++ apps/hindsight-service/core/services/notifications/membership_change_flow.py
@@ -1,0 +1,394 @@
+"""
+MembershipChangeFlow: handles notifications for organization membership changes
+(added, role changed, removed).
+
+Nested closures from the original notify_membership_added, notify_role_changed,
+and notify_membership_removed have been promoted to private methods on this class.
+"""
+
+import uuid
+import threading
+from datetime import datetime
+from typing import Optional, Dict, Any
+
+from core.services.notifications.dispatcher import NotificationDispatcher
+from core.services.notifications.constants import (
+    EVENT_ORG_MEMBERSHIP_ADDED,
+    EVENT_ORG_MEMBERSHIP_REMOVED,
+    EVENT_ORG_ROLE_CHANGED,
+    TEMPLATE_MEMBERSHIP_ADDED,
+    TEMPLATE_MEMBERSHIP_REMOVED,
+    TEMPLATE_ROLE_CHANGED,
+)
+
+
+class MembershipChangeFlow:
+    """
+    Handles all notification flows for organization membership changes.
+    """
+
+    def __init__(self, dispatcher: NotificationDispatcher, email_service=None):
+        self._dispatcher = dispatcher
+        self.email_service = email_service
+
+    # ------------------------------------------------------------------ #
+    # Private helpers (promoted from nested closures)                     #
+    # ------------------------------------------------------------------ #
+
+    def _update_email_status(
+        self,
+        email_log_id: uuid.UUID,
+        status: str,
+        message_id: Optional[str] = None,
+        error: Optional[str] = None,
+    ) -> None:
+        """Thin wrapper delegating to dispatcher._safe_update_email_status."""
+        self._dispatcher._safe_update_email_status(
+            email_log_id,
+            status,
+            provider_message_id=message_id,
+            error_message=error,
+        )
+
+    def _send_email_for_membership_added(
+        self,
+        email_log_id: uuid.UUID,
+        email_address: str,
+        subject: str,
+        html: str,
+        text: str,
+    ) -> None:
+        """
+        Background-thread sender for notify_membership_added.
+        Promoted from the nested _send closure in the original implementation.
+        """
+        import asyncio
+        try:
+            send_res = asyncio.run(self.email_service.send_email(
+                to_email=email_address,
+                subject=subject,
+                html_content=html,
+                text_content=text
+            ))
+            if send_res.get('success'):
+                self._update_email_status(email_log_id, 'sent', message_id=send_res.get('message_id'))
+            else:
+                self._update_email_status(email_log_id, 'failed', error=send_res.get('error'))
+        except Exception as e:
+            self._update_email_status(email_log_id, 'failed', error=str(e))
+
+    def _send_email_for_role_changed(
+        self,
+        email_log_id: uuid.UUID,
+        email_address: str,
+        subject: str,
+        html: str,
+        text: str,
+    ) -> None:
+        """
+        Background-thread sender for notify_role_changed.
+        Promoted from the nested _send closure in the original implementation.
+        """
+        import asyncio
+        try:
+            send_res = asyncio.run(self.email_service.send_email(
+                to_email=email_address,
+                subject=subject,
+                html_content=html,
+                text_content=text
+            ))
+            if send_res.get('success'):
+                self._dispatcher._safe_update_email_status(
+                    email_log_id, 'sent', provider_message_id=send_res.get('message_id')
+                )
+            else:
+                self._dispatcher._safe_update_email_status(
+                    email_log_id, 'failed', error_message=send_res.get('error')
+                )
+        except Exception as e:
+            self._dispatcher._safe_update_email_status(email_log_id, 'failed', error_message=str(e))
+
+    def _send_email_for_membership_removed(
+        self,
+        email_log_id: uuid.UUID,
+        email_address: str,
+        subject: str,
+        html: str,
+        text: str,
+    ) -> None:
+        """
+        Background-thread sender for notify_membership_removed.
+        Promoted from the nested _send closure in the original implementation.
+        """
+        import asyncio
+        try:
+            send_res = asyncio.run(self.email_service.send_email(
+                to_email=email_address,
+                subject=subject,
+                html_content=html,
+                text_content=text
+            ))
+            if send_res.get('success'):
+                self._dispatcher._safe_update_email_status(
+                    email_log_id, 'sent', provider_message_id=send_res.get('message_id')
+                )
+            else:
+                self._dispatcher._safe_update_email_status(
+                    email_log_id, 'failed', error_message=send_res.get('error')
+                )
+        except Exception as e:
+            self._dispatcher._safe_update_email_status(email_log_id, 'failed', error_message=str(e))
+
+    # ------------------------------------------------------------------ #
+    # Public flow methods                                                 #
+    # ------------------------------------------------------------------ #
+
+    def notify_membership_added(
+        self,
+        user_id: uuid.UUID,
+        user_email: str,
+        organization_name: str,
+        role: str,
+        added_by_name: str,
+        organization_id: Optional[uuid.UUID] = None,
+        added_by_user_id: Optional[uuid.UUID] = None,
+    ) -> Dict[str, Any]:
+        """
+        Send notifications when a user is added to an organization.
+        """
+        preferences = self._dispatcher.get_user_preferences(user_id)
+        membership_prefs = preferences.get('org_membership_added', {'email_enabled': True, 'in_app_enabled': True})
+
+        result = {}
+
+        # Create in-app notification if enabled
+        if membership_prefs['in_app_enabled']:
+            metadata: Dict[str, Any] = {
+                'organization_name': organization_name,
+                'role': role,
+                'added_by_name': added_by_name,
+            }
+            if organization_id:
+                metadata['organization_id'] = str(organization_id)
+            if added_by_user_id:
+                metadata['added_by_user_id'] = str(added_by_user_id)
+
+            # Title expected by tests includes trailing exclamation mark
+            notification = self._dispatcher.create_notification(
+                user_id=user_id,
+                # Use canonical event_type constant for stored notifications
+                event_type=EVENT_ORG_MEMBERSHIP_ADDED,
+                title=f"Welcome to {organization_name}!",
+                # Use wording that includes the exact substring tests look for
+                message=f"{added_by_name} added you to the organization {organization_name} as a {role}.",
+                metadata=metadata
+            )
+            result['in_app_notification'] = notification
+
+        # Create email notification log if email enabled
+        if membership_prefs['email_enabled']:
+            email_log = self._dispatcher.create_email_notification_log(
+                notification_id=result.get('in_app_notification', {}).id if 'in_app_notification' in result else None,
+                user_id=user_id,
+                # use canonical event type for logs
+                event_type=EVENT_ORG_MEMBERSHIP_ADDED,
+                email_address=user_email,
+                subject=f"Welcome to {organization_name}"
+            )
+            result['email_log'] = email_log
+
+            # Send the email in background so API response isn't blocked.
+            try:
+                if not self.email_service:
+                    # If no email service (rare in tests), mark as failed but continue
+                    self._dispatcher.update_email_status(
+                        email_log.id, 'failed', error_message='No email service configured'
+                    )
+                else:
+                    from core.utils.urls import get_app_base_url
+                    base_url = get_app_base_url()
+                    # Build context expected by tests
+                    context = {
+                        'user_name': (user_email.split('@')[0] if user_email else ''),
+                        'organization_name': organization_name,
+                        'invited_by': added_by_name,
+                        'role': role,
+                        'dashboard_url': f'{base_url}/dashboard'
+                    }
+
+                    # Render template synchronously (so patched render_template gets called)
+                    try:
+                        html, text = self.email_service.render_template(TEMPLATE_MEMBERSHIP_ADDED, context)
+                    except Exception as e:
+                        # If rendering fails, mark email log and return
+                        self._dispatcher.update_email_status(
+                            email_log.id, 'failed', error_message=f'Template render failed: {str(e)}'
+                        )
+                        return result
+
+                    # Send in background thread using promoted private method
+                    t = threading.Thread(
+                        target=self._send_email_for_membership_added,
+                        args=(email_log.id, email_log.email_address, email_log.subject, html, text),
+                        daemon=True,
+                    )
+                    t.start()
+
+                    result['email_result'] = {'dispatched_in_background': True}
+            except Exception as e:
+                self._dispatcher._safe_update_email_status(
+                    email_log.id, 'failed', error_message=f"Email dispatch error: {str(e)}"
+                )
+
+        return result
+
+    def notify_role_changed(
+        self,
+        user_id: uuid.UUID,
+        user_email: str,
+        organization_name: str,
+        old_role: str,
+        new_role: str,
+        changed_by_name: str
+    ) -> Dict[str, Any]:
+        """
+        Notify a user that their role in an organization has changed.
+        """
+        prefs = self._dispatcher.get_user_preferences(user_id).get(
+            'org_role_changed', {'email_enabled': True, 'in_app_enabled': True}
+        )
+        result: Dict[str, Any] = {}
+
+        title = f"Role changed in {organization_name}"
+        message = f"Your role in {organization_name} changed from {old_role} to {new_role} by {changed_by_name}."
+
+        if prefs.get('in_app_enabled'):
+            n = self._dispatcher.create_notification(
+                user_id=user_id,
+                event_type=EVENT_ORG_ROLE_CHANGED,
+                title=title,
+                message=message,
+                metadata={
+                    'organization_name': organization_name,
+                    'old_role': old_role,
+                    'new_role': new_role,
+                    'changed_by': changed_by_name,
+                }
+            )
+            result['in_app_notification'] = n
+
+        if prefs.get('email_enabled'):
+            log = self._dispatcher.create_email_notification_log(
+                notification_id=result.get('in_app_notification', {}).id if 'in_app_notification' in result else None,
+                user_id=user_id,
+                email_address=user_email,
+                event_type=EVENT_ORG_ROLE_CHANGED,
+                subject=f"Your role changed in {organization_name}"
+            )
+            result['email_log'] = log
+            try:
+                template_context = {
+                    'organization_name': organization_name,
+                    'old_role': old_role,
+                    'new_role': new_role,
+                    'changed_by_name': changed_by_name,
+                    'date_changed': datetime.now().strftime('%B %d, %Y'),
+                    'current_year': datetime.now().year,
+                }
+
+                # Render template synchronously
+                try:
+                    html, text = self.email_service.render_template(TEMPLATE_ROLE_CHANGED, template_context)
+                except Exception as e:
+                    # If rendering fails, mark email log and return
+                    self._dispatcher.update_email_status(
+                        log.id, 'failed', error_message=f'Template render failed: {str(e)}'
+                    )
+                    return result
+
+                # Send in background thread using promoted private method
+                t = threading.Thread(
+                    target=self._send_email_for_role_changed,
+                    args=(log.id, log.email_address, log.subject, html, text),
+                    daemon=True,
+                )
+                t.start()
+
+                result['email_result'] = {'dispatched_in_background': True}
+            except Exception as e:
+                self._dispatcher._safe_update_email_status(
+                    log.id, 'failed', error_message=f"Email dispatch error: {str(e)}"
+                )
+
+        return result
+
+    def notify_membership_removed(
+        self,
+        user_id: uuid.UUID,
+        user_email: str,
+        organization_name: str,
+        removed_by_name: str
+    ) -> Dict[str, Any]:
+        """
+        Notify a user that they have been removed from an organization.
+        """
+        prefs = self._dispatcher.get_user_preferences(user_id).get(
+            'org_membership_removed', {'email_enabled': True, 'in_app_enabled': True}
+        )
+        result: Dict[str, Any] = {}
+
+        title = f"Removed from {organization_name}"
+        message = f"You have been removed from {organization_name} by {removed_by_name}."
+
+        if prefs.get('in_app_enabled'):
+            n = self._dispatcher.create_notification(
+                user_id=user_id,
+                event_type=EVENT_ORG_MEMBERSHIP_REMOVED,
+                title=title,
+                message=message,
+                metadata={'organization_name': organization_name, 'removed_by': removed_by_name}
+            )
+            result['in_app_notification'] = n
+
+        if prefs.get('email_enabled'):
+            log = self._dispatcher.create_email_notification_log(
+                notification_id=result.get('in_app_notification', {}).id if 'in_app_notification' in result else None,
+                user_id=user_id,
+                email_address=user_email,
+                event_type=EVENT_ORG_MEMBERSHIP_REMOVED,
+                subject=f"You were removed from {organization_name}"
+            )
+            result['email_log'] = log
+            try:
+                template_context = {
+                    'organization_name': organization_name,
+                    'removed_by_name': removed_by_name,
+                    'date_removed': datetime.now().strftime('%B %d, %Y'),
+                    'current_year': datetime.now().year,
+                }
+
+                # Render template synchronously
+                try:
+                    html, text = self.email_service.render_template(TEMPLATE_MEMBERSHIP_REMOVED, template_context)
+                except Exception as e:
+                    # If rendering fails, mark email log and return
+                    self._dispatcher.update_email_status(
+                        log.id, 'failed', error_message=f'Template render failed: {str(e)}'
+                    )
+                    return result
+
+                # Send in background thread using promoted private method
+                t = threading.Thread(
+                    target=self._send_email_for_membership_removed,
+                    args=(log.id, log.email_address, log.subject, html, text),
+                    daemon=True,
+                )
+                t.start()
+
+                result['email_result'] = {'dispatched_in_background': True}
+            except Exception as e:
+                self._dispatcher._safe_update_email_status(
+                    log.id, 'failed', error_message=f"Email dispatch error: {str(e)}"
+                )
+
+        return result
